Add limitation of http connection for MMS server.
